Move tip logic from AppState to TipProvider
