+++ Bemol/AppEnvironment.swift
@@ -22,6 +22,7 @@ import os
 struct AppEnvironment {
   let notePlayer: NotePlayer
   let practiceManager: PracticeManager
+  let tipProvider: TipProvider
   let preferences: Preferences
   let logger: Logger
 }
